+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Added Collection Detail to Layout Kit([#893](https://github.com/powerhome/playbook/pull/893) @evalouderback)
+
 ## [v5.1.0] 2020-6-26
 
 ### Added