+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Added New Layout Kit (Content) for Rails and React ([#894](https://github.com/powerhome/playbook/pull/894) @megantrimble)
+
 ## [v5.1.0] 2020-6-26
 
 ### Added