+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- Added Collection Detail to Layout Kit([#893](https://github.com/powerhome/playbook/pull/893) @evalouderback)
 - Added New Layout Kit (Content) for Rails and React ([#894](https://github.com/powerhome/playbook/pull/894) @megantrimble)
 
 ## [v5.1.0] 2020-6-26